fixed PUT method in notes

+++ api/resources/note.py
@@ -9,7 +9,7 @@ from webargs import fields
 @doc(tags=['Notes'])
 class NoteResource(MethodResource):
     @auth.login_required
-    @doc(description='Get notes by user id')
+    @doc(description='Get notes by user id', summary="Get notes")
     @marshal_with(NoteSchema)
     def get(self, note_id):
         author = g.user
@@ -20,32 +20,37 @@ class NoteResource(MethodResource):
             abort(403, error=f"Forbidden")
         return note, 200
 
-    # FIXME Fix put like put in users
     @auth.login_required
-    @doc(security=[{"basicAuth": []}])
+    @doc(security=[{"basicAuth": []}], description='Edit users notes by note id', summary="Edit notes")
     @marshal_with(NoteSchema)
-    def put(self, note_id):
+    @use_kwargs({"text": fields.Str(), "private": fields.Boolean()})
+    def put(self, note_id, **kwargs):
         author = g.user
-        parser = reqparse.RequestParser()
-        parser.add_argument("text", required=True)
-        parser.add_argument("private", type=bool)  # чтобы не передовался тип приватности в виде строки
-        note_data = parser.parse_args()
+        #parser = reqparse.RequestParser()
+        #parser.add_argument("text", required=True)
+        #parser.add_argument("private", type=bool)  # чтобы не передовался тип приватности в виде строки
+        #note_data = parser.parse_args()
         note = NoteModel.query.get(note_id)
+        note.text = kwargs["text"]
+        note.private = kwargs["private"]
         if not note:
             abort(404, error=f"note {note_id} not found")
         if note.author != author:
             abort(403, error=f"Forbidden")
-        note.text = note_data["text"]
-        note.private = note_data.get("private") or note.private
+        #note.text = note_data["text"]
+        #note.private = note_data.get("private") or note.private
         note.save()
         return note, 200
 
     @auth.login_required
-    @doc(security=[{"basicAuth": []}])
+    @doc(security=[{"basicAuth": []}], description='Delete users notes by note id', summary="Delete notes")
     def delete(self, note_id):
+        author = g.user
         note = NoteModel.query.get(note_id)
         if not note:
             abort(404, error=f"Note with id:{note_id} not found")
+        if note.author != author:
+            abort(403, error=f"Forbidden")
         note_dict = note_schema.dump(note)
         note.delete()
         return note_dict, 200
@@ -66,7 +71,7 @@ class NotesListResource(MethodResource):
     @marshal_with(NoteSchema)
     def post(self, **kwargs):
         author = g.user
-        #parser = reqparse.RequestParser()
+        #parser = reqparse.RequestParser()      Заменил парсер reqparse на использование декоратора use_kwargs
         #parser.add_argument("text", required=True)
         #parser.add_argument("private", type=bool)
         #note_data = parser.parse_args()
